Update README.md
This project aims to develop Ekhata. The project titled "Ekhata" is a web-based application. It was created using PHP and MySQL. This project has tried to incorporate all the advanced features of PHP such as to meet the project goals. A necessary platform for ordinary shopkeepers or small traders. This solution helps shopkeepers or vendors to keep all records of the product price, sales, purchase information, customer information & payment information and also show daily or monthly profit. This software solution helps increase business growth and smart business management.